Рефакторинг приложения. Улучшение структуры разбиения по слоям.
