+++ app/src/main/java/ru/marslab/filmoteca/domain/Store.kt
@@ -1,0 +1,16 @@
+package ru.marslab.filmoteca.domain
+
+import ru.marslab.filmoteca.BuildConfig
+import ru.marslab.filmoteca.domain.model.RequestToken
+
+object Store {
+    var guestSessionId: String? = null
+    var sessionId: String? = null
+    var requestToken: RequestToken? = null
+
+    val apiKeyV3: String
+        get() = BuildConfig.API_KEY_V3
+    val apiKeyV4: String
+        get() = BuildConfig.API_KEY_V4
+
+}
