+++ app/src/main/java/ru/marslab/filmoteca/domain/di/DataModule.kt
@@ -10,18 +10,11 @@ import dagger.hilt.android.components.ViewModelComponent
 import dagger.hilt.android.scopes.ViewModelScoped
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
-import ru.marslab.filmoteca.data.RepositoryNetworkImpl
 import ru.marslab.filmoteca.data.retrofit.MovieApi
-import ru.marslab.filmoteca.domain.Repository
-import javax.inject.Singleton
 
 @Module
 @InstallIn(ViewModelComponent::class)
 object DataModule {
-
-    @ViewModelScoped
-    @Provides
-    fun provideRepository(api: MovieApi): Repository = RepositoryNetworkImpl(api)
 
     @ViewModelScoped
     @Provides
