+++ app/src/main/java/ru/marslab/filmoteca/domain/repository/MovieRepository.kt
@@ -1,0 +1,7 @@
+package ru.marslab.filmoteca.domain.repository
+
+import ru.marslab.filmoteca.domain.model.Movie
+
+interface MovieRepository {
+    suspend fun getMovieDetail(id: Int): Movie?
+}
