+++ app/src/main/java/ru/marslab/filmoteca/domain/repository/Repository.kt
@@ -1,0 +1,8 @@
+package ru.marslab.filmoteca.domain.repository
+
+import ru.marslab.filmoteca.domain.model.RequestToken
+
+interface Repository {
+
+    suspend fun createRequestToken(): RequestToken?
+}
