+++ app/src/main/java/ru/marslab/filmoteca/ui/movie/MainFilmViewModel.kt
@@ -2,7 +2,7 @@ package ru.marslab.filmoteca.ui.movie
 
 import androidx.lifecycle.ViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
-import ru.marslab.filmoteca.domain.Repository
+import ru.marslab.filmoteca.domain.repository.Repository
 import javax.inject.Inject
 
 @HiltViewModel
