+++ app/src/main/java/ru/marslab/filmoteca/ui/movie/MovieDetailViewModel.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
-import ru.marslab.filmoteca.domain.Repository
+import ru.marslab.filmoteca.domain.repository.Repository
 import ru.marslab.filmoteca.ui.model.MovieDetailUi
 import javax.inject.Inject
